requirements.txt
streamlit
Pillow
torch
diffusers
transformers
groq
python-dotenv